Initial commit
The main work is in try_2 which is currently wrong because I'm trying to transfer the learning to our classification for steam. You can find the details on how to do it (I hope so) on 'Transfer_learning_with_MobileNet_v1.ipynb'. I tried to split the data to create the training and test set. The resize.py document would resize the whole file but I don't actively use it now as mobilenet resizes independently. Dataset 2 only has 3 items to test the model on quickly. When I first downloaded the images a couple of them were corrupted (idk why) so I took them out from the dataset.json file, it's just 2 images.